+++ wayland-scanner/tests/scanner_assets/test-server-code.rs
@@ -11,7 +11,7 @@ pub mod wl_callback {
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_DONE_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_DONE_OPCODE: u32 = 0u32;
+    pub const EVT_DONE_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {}
@@ -129,43 +129,43 @@ pub mod test_global {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_MANY_ARGS_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_MANY_ARGS_OPCODE: u32 = 0u32;
+    pub const REQ_MANY_ARGS_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_GET_SECONDARY_SINCE: u32 = 2u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_GET_SECONDARY_OPCODE: u32 = 1u32;
+    pub const REQ_GET_SECONDARY_OPCODE: u16 = 1u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_GET_TERTIARY_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_GET_TERTIARY_OPCODE: u32 = 2u32;
+    pub const REQ_GET_TERTIARY_OPCODE: u16 = 2u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_LINK_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_LINK_OPCODE: u32 = 3u32;
+    pub const REQ_LINK_OPCODE: u16 = 3u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 4u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 4u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 4u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_REVERSE_LINK_SINCE: u32 = 5u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_REVERSE_LINK_OPCODE: u32 = 5u32;
+    pub const REQ_REVERSE_LINK_OPCODE: u16 = 5u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_NEWID_AND_ALLOW_NULL_SINCE: u32 = 5u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_NEWID_AND_ALLOW_NULL_OPCODE: u32 = 6u32;
+    pub const REQ_NEWID_AND_ALLOW_NULL_OPCODE: u16 = 6u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_MANY_ARGS_EVT_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_MANY_ARGS_EVT_OPCODE: u32 = 0u32;
+    pub const EVT_MANY_ARGS_EVT_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_ACK_SECONDARY_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_ACK_SECONDARY_OPCODE: u32 = 1u32;
+    pub const EVT_ACK_SECONDARY_OPCODE: u16 = 1u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_CYCLE_QUAD_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_CYCLE_QUAD_OPCODE: u32 = 2u32;
+    pub const EVT_CYCLE_QUAD_OPCODE: u16 = 2u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -621,7 +621,7 @@ pub mod secondary {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 2u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 0u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -731,7 +731,7 @@ pub mod tertiary {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 0u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -841,7 +841,7 @@ pub mod quad {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 0u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
